upgrade to latest workflow
