+++ .gitattributes
@@ -1,2 +1,3 @@
 
-yarn.lock -diff
+*.gltf -diff linguist-generated
+yarn.lock -diff linguist-generated
