+++ example/main.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+import ReactDOM from "react-dom";
+import Container from "./container";
+
+let mountTarget = document.querySelector(".app");
+
+ReactDOM.render(<Container />, mountTarget);
