+++ package.json
@@ -4,53 +4,70 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "webpack-dev-server --hot",
-    "dev": "webpack-dev-server --hot",
-    "release": "rm -rf dist/* && webpack --progress --config webpack.release.js && cp -v data/*.gltf dist/",
-    "upload": "rsync -avr --progress dist/ fe.jimu.io:/web-assets/repo/jimengio/try-three-model",
+    "dev": "cp -vr data webpack/ && webpack-dev-server --hot --config webpack/dev.js",
+    "watch": "cp -vr data webpack/ && webpack-dev-server --hot --config webpack/dev.js",
+    "dll": "rm -rfv webpack/dll/* && time webpack --config webpack/dll-dev.js",
+    "dll-release": "time webpack --config webpack/dll-release.js",
+    "release": "rm -rfv dist/* && yarn dll-release && time webpack --config webpack/release.js && cp -v data/* dist/",
     "build": "yarn release",
-    "serve": "http-server dist/ -s"
+    "serve": "http-server dist/ -s",
+    "compile": "rm -rfv lib/* && tsc -d --project tsconfig-compile.json --outDir lib/",
+    "upload": "rsync -avr --progress dist/ fe.jimu.io:/web-assets/repo/jimengio/try-three-model",
+    "gen-router": "ts-node -P tsconfig-node.json -r tsconfig-paths/register example/gen-router.ts",
+    "postinstall": "rm -rfv node_modules/@types/*/node_modules/@types/react"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@types/history": "^4.5.2",
-    "@types/lodash": "^4.14.80",
-    "@types/react": "^16.0.18",
-    "@types/react-dom": "^16.0.2",
-    "@types/react-router-dom": "^4.2.0",
-    "@types/three": "^0.92.0",
-    "awesome-typescript-loader": "^3.2.3",
-    "css-loader": "^0.28.4",
-    "emotion": "^8.0.9",
-    "eventemitter3": "^2.0.3",
-    "extract-text-webpack-plugin": "^3.0.2",
-    "file-loader": "^1.1.5",
+    "@jimengio/ci-progress-webpack-plugin": "^0.1.0",
+    "@jimengio/doc-frame": "^0.1.11",
+    "@jimengio/flex-styles": "^0.1.6",
+    "@jimengio/router-code-generator": "^0.2.5",
+    "@jimengio/shared-utils": "^0.1.5",
+    "@types/node": "^13.7.0",
+    "@types/prettier": "^1.19.0",
+    "@types/query-string": "^6.3.0",
+    "@types/react": "^16.9.19",
+    "@types/react-dom": "^16.9.5",
+    "cache-loader": "^4.1.0",
+    "css-loader": "^3.4.2",
+    "duplicate-package-checker-webpack-plugin": "^3.0.0",
+    "file-loader": "^5.0.2",
+    "fork-ts-checker-webpack-plugin": "^4.0.3",
+    "html-webpack-plugin": "^3.2.0",
+    "html-webpack-tags-plugin": "^2.0.17",
+    "http-server": "^0.12.1",
+    "mini-css-extract-plugin": "^0.9.0",
+    "prettier": "^1.19.1",
+    "style-loader": "^1.1.3",
+    "thread-loader": "^2.1.3",
+    "ts-loader": "^6.2.1",
+    "ts-node": "^8.6.2",
+    "tsconfig-paths": "^3.9.0",
+    "tsconfig-paths-webpack-plugin": "^3.2.0",
+    "typescript": "^3.7.5",
+    "typescript-styled-plugin": "^0.15.0",
+    "url-loader": "^3.0.0",
+    "webpack": "^4.41.5",
+    "webpack-bundle-analyzer": "^3.6.0",
+    "webpack-cli": "^3.3.10",
+    "webpack-dev-server": "^3.10.2",
+    "dayjs": "^1.8.19",
+    "@jimengio/ruled-router": "^0.2.23",
+    "emotion": "^10.0.27",
     "font-awesome": "^4.7.0",
-    "history": "^4.6.1",
-    "html-webpack-plugin": "^2.30.1",
-    "http-server": "^0.10.0",
-    "lodash": "^4.17.4",
-    "mobx": "^3.1.16",
-    "mobx-react": "^4.2.1",
-    "mobx-state-tree": "^1.1.0",
-    "moment": "^2.19.1",
-    "react": "^16.0.0",
-    "react-dom": "^16.0.0",
-    "react-router": "^4.1.1",
-    "react-router-dom": "^4.1.1",
-    "shortid": "^2.2.8",
-    "style-loader": "^0.19.0",
-    "typescript": "^2.6.1",
-    "typescript-styled-plugin": "^0.1.2",
-    "uglifyjs-webpack-plugin": "^1.0.1",
-    "url-loader": "^0.6.2",
-    "webpack": "^3.8.1",
-    "webpack-dev-server": "^2.9.3",
-    "webpack-hud": "^0.1.1"
+    "immer": "^5.3.2",
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0",
+    "query-string": "^6.10.1",
+    "webpack-hud": "^0.1.2"
+  },
+  "peerDependencies": {
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0"
   },
   "dependencies": {
-    "three": "^0.92.0"
+    "three": "^0.113.2"
   }
 }
