+++ webpack/data/.gitattributes
@@ -1,0 +1,2 @@
+
+*.gltf -diff linguist-generated
